수익노트 수정 response 값 중 sellCount 값이 int형으로 내려지도록 수정

+++ income_note/views.py
@@ -124,7 +124,7 @@ class IncomeNoteAPI(APIView):
             sellDate = request.data.get("sellDate")
             purchasePrice = Decimal(request.data.get("purchasePrice"))
             sellPrice = Decimal(request.data.get("sellPrice"))
-            sellCount = Decimal(request.data.get("sellCount"))
+            sellCount = int(request.data.get("sellCount"))
 
             # 계산
             realPainLossesAmount = (sellPrice - purchasePrice) * sellCount
